requirements.txt
pyyaml
pypresence